Enable multi-tests for CPU variants
